// src/app/models/value-types.ts
export enum ValueTypes {
  BuyDate = 'buyDate',
  BuyCurrency = 'buyCurrency',
  BuyRate = 'buyRate',
  BuyPrice = 'buyPrice',
  BuyCurrencyExchanged = 'buyCurrencyExchanged',
  BuyRateExchanged = 'buyRateExchanged',

  SellDate = 'sellDate',
  SellCurrency = 'sellCurrency',
  SellRate = 'sellRate',
  SellPrice = 'sellPrice',
  SellCurrencyExchanged = 'sellCurrencyExchanged',
  SellRateExchanged = 'sellRateExchanged',
}